LunchtimeFXOS: Add debug version and include jquery. Improve gulpfile
There is now an index.debug.html, which includes non-minified js with
source maps, as well as a livereload script (so that the page is reloaded
in the browser if you save a css file, for instance). index.html includes
minified bundle.min.js and should start significantly faster on a slow
device.

The browserify shim stuff that I previously used for jquery (to load it
from a CDN that had a greater chance of existing in browser cache) didn't
work on the Peak, so I just included it in the bundle..